pyproject.toml
[project]
name = "coherence-engine"
version = "0.1.0"
description = "Lightweight Data Coherence Engine (FastAPI)"
readme = "README.md"
requires-python = ">=3.11"
authors = [{name="Sheldon H. Gordon"}]

dependencies = [
  "fastapi==0.115.0",
  "uvicorn[standard]==0.30.6",
  "httpx==0.27.2",
  "pydantic==2.9.2",
  "python-dotenv==1.0.1",
  "pandas==2.2.3",
  "numpy==2.1.2",
  "streamlit==1.39.0",
]

[tool.pytest.ini_options]
addopts = "-q"
pythonpath = ["app"]
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.black]
line-length = 100
target-version = ["py311"]
